requirements.txt
discord.py
apscheduler
python-dotenv